Conflicts:
	.travis.yml
	CONTRIBUTING.md
	README.md
	app/Http/Controllers/AccountController.php
	app/Http/Controllers/AppController.php
	app/Http/Controllers/BaseAPIController.php
	app/Http/Controllers/CreditController.php
	app/Http/Controllers/DashboardController.php
	app/Http/Controllers/DocumentController.php
	app/Http/Controllers/PaymentController.php
	app/Http/Controllers/QuoteApiController.php
	app/Http/Requests/EntityRequest.php
	app/Http/routes.php
	app/Listeners/AnalyticsListener.php
	app/Models/Account.php
	app/Models/Credit.php
	app/Models/Document.php
	app/Models/EntityModel.php
	app/Models/Expense.php
	app/Models/Invoice.php
	app/Models/Task.php
	app/Models/VendorContact.php
	app/Ninja/Notifications/PushFactory.php
	app/Ninja/Repositories/DocumentRepository.php
	app/Ninja/Repositories/ExpenseRepository.php
	app/Ninja/Repositories/InvoiceRepository.php
	app/Services/PushService.php
	composer.json
	composer.lock
	database/migrations/2014_05_17_175626_add_quotes.php
	database/seeds/UserTableSeeder.php
	resources/lang/en/texts.php
	resources/views/accounts/customize_design.blade.php
	resources/views/accounts/invoice_design.blade.php
	resources/views/accounts/management.blade.php
	resources/views/expenses/edit.blade.php
	resources/views/header.blade.php
	resources/views/invoices/edit.blade.php
	resources/views/master.blade.php
	resources/views/payments/payment.blade.php
	resources/views/users/edit.blade.php

+++ app/Ninja/Repositories/ExpenseRepository.php
@@ -6,8 +6,6 @@ use Auth;
 use App\Models\Expense;
 use App\Models\Vendor;
 use App\Models\Document;
-use App\Ninja\Repositories\BaseRepository;
-use Session;
 
 class ExpenseRepository extends BaseRepository
 {
@@ -36,22 +34,8 @@ class ExpenseRepository extends BaseRepository
     public function findVendor($vendorPublicId)
     {
         $vendorId = Vendor::getPrivateId($vendorPublicId);
-        $accountid = \Auth::user()->account_id;
-        $query = DB::table('expenses')
-                    ->join('accounts', 'accounts.id', '=', 'expenses.account_id')
-                    ->where('expenses.account_id', '=', $accountid)
-                    ->where('expenses.vendor_id', '=', $vendorId)
-                    ->select(
-                        'expenses.id',
-                        'expenses.expense_date',
-                        'expenses.amount',
-                        'expenses.public_notes',
-                        'expenses.public_id',
-                        'expenses.deleted_at',
-                        'expenses.should_be_invoiced',
-                        'expenses.created_at',
-                        'expenses.user_id'
-                    );
+
+        $query = $this->find()->where('expenses.vendor_id', '=', $vendorId);
 
         return $query;
     }
@@ -65,6 +49,7 @@ class ExpenseRepository extends BaseRepository
                     ->leftJoin('contacts', 'contacts.client_id', '=', 'clients.id')
                     ->leftjoin('vendors', 'vendors.id', '=', 'expenses.vendor_id')
                     ->leftJoin('invoices', 'invoices.id', '=', 'expenses.invoice_id')
+                    ->leftJoin('expense_categories', 'expenses.expense_category_id', '=', 'expense_categories.id')
                     ->where('expenses.account_id', '=', $accountid)
                     ->where('contacts.deleted_at', '=', null)
                     ->where('vendors.deleted_at', '=', null)
@@ -91,8 +76,10 @@ class ExpenseRepository extends BaseRepository
                         'expenses.expense_currency_id',
                         'expenses.invoice_currency_id',
                         'expenses.user_id',
+                        'expense_categories.name as category',
                         'invoices.public_id as invoice_public_id',
                         'invoices.user_id as invoice_user_id',
+                        'invoices.balance',
                         'vendors.name as vendor_name',
                         'vendors.public_id as vendor_public_id',
                         'vendors.user_id as vendor_user_id',
@@ -115,7 +102,8 @@ class ExpenseRepository extends BaseRepository
             $query->where(function ($query) use ($filter) {
                 $query->where('expenses.public_notes', 'like', '%'.$filter.'%')
                       ->orWhere('clients.name', 'like', '%'.$filter.'%')
-                      ->orWhere('vendors.name', 'like', '%'.$filter.'%');
+                      ->orWhere('vendors.name', 'like', '%'.$filter.'%')
+                      ->orWhere('expense_categories.name', 'like', '%'.$filter.'%');;
             });
         }
 
@@ -144,7 +132,7 @@ class ExpenseRepository extends BaseRepository
             $expense->private_notes = trim($input['private_notes']);
         }
         $expense->public_notes = trim($input['public_notes']);
-        $expense->should_be_invoiced = isset($input['should_be_invoiced']) || $expense->client_id ? true : false;
+        $expense->should_be_invoiced = isset($input['should_be_invoiced']) && floatval($input['should_be_invoiced']) || $expense->client_id ? true : false;
 
         if ( ! $expense->expense_currency_id) {
             $expense->expense_currency_id = \Auth::user()->account->getCurrencyId();
@@ -160,7 +148,7 @@ class ExpenseRepository extends BaseRepository
         $expense->save();
 
         // Documents
-        $document_ids = !empty($input['document_ids'])?array_map('intval', $input['document_ids']):array();;
+        $document_ids = !empty($input['document_ids'])?array_map('intval', $input['document_ids']):[];;
         foreach ($document_ids as $document_id){
             // check document completed upload before user submitted form
             if ($document_id) {
@@ -173,29 +161,13 @@ class ExpenseRepository extends BaseRepository
             }
         }
 
-        if(!empty($input['documents']) && Auth::user()->can('create', ENTITY_DOCUMENT)){
-            // Fallback upload
-            $doc_errors = array();
-            foreach($input['documents'] as $upload){
-                $result = $this->documentRepo->upload($upload);
-                if(is_string($result)){
-                    $doc_errors[] = $result;
+        // prevent loading all of the documents if we don't have to
+        if ( ! $expense->wasRecentlyCreated) {
+            foreach ($expense->documents as $document){
+                if ( ! in_array($document->public_id, $document_ids)){
+                    // Not checking permissions; deleting a document is just editing the invoice
+                    $document->delete();
                 }
-                else{
-                    $result->expense_id = $expense->id;
-                    $result->save();
-                    $document_ids[] = $result->public_id;
-                }
-            }
-            if(!empty($doc_errors)){
-                Session::flash('error', implode('<br>',array_map('htmlentities',$doc_errors)));
-            }
-        }
-
-        foreach ($expense->documents as $document){
-            if(!in_array($document->public_id, $document_ids)){
-                // Not checking permissions; deleting a document is just editing the invoice
-                $document->delete();
             }
         }
